requirement.txt
Babel==2.6.0
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
click==6.7
cymysql==0.9.12
Flask==1.0.2
Flask-Babel==0.12.0
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
python-dotenv==0.9.1
pytz==2018.6
raven==6.9.0
requests==2.20.0
SQLAlchemy==1.2.10
urllib3==1.24
Werkzeug==0.14.1
